DAY_5/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.2.0
django-use-email-as-username==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-excel==2.4.0
drf-renderer-xlsx==1.0.0
et-xmlfile==1.1.0
openpyxl==3.1.2
Pillow==10.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
XlsxWriter==3.1.6
xlwt==1.3.0
